requirements.txt
browser-use
playwright
openai
langchain-openai
python-dotenv